Improved Filter Screen UI. Handled Location Off case

+++ app/src/main/java/com/example/nearbyplaces/GetFilterValues.java
@@ -1,5 +1,5 @@
 package com.example.nearbyplaces;
 
 public interface GetFilterValues {
-    void getFilterValues(String type, Double radius);
+    void getFilterValues(String type, Float radius);
 }
